requirements.txt
certifi==2020.12.5
chardet==4.0.0
getmac==0.8.2
gitdb==4.0.6
GitPython==3.1.14
idna==2.10
numpy==1.20.1
opencv-python==4.5.1.48
paho-mqtt==1.5.1
playsound==1.2.2
PyBluez==0.23
python-decouple==3.4
requests==2.25.1
smmap==3.0.5
urllib3==1.26.4

pip~=21.0.1
Jinja2~=2.11.3
Werkzeug~=1.0.1
itsdangerous~=1.1.0
click~=7.1.2
MarkupSafe~=1.1.1
setuptools~=54.2.0
Flask~=1.1.2
gTTS~=2.2.2